Feature/vinhkt2/events management (#26)
* add paging and filter, change table event fields to nullable

* add updateasync

* update add api

* fix comments

* fix comments

* remove unuse usings

+++ EventTicketingManagementSystem/EventTicketingManagementSystem/Services/Implements/EventService.cs
@@ -1,6 +1,8 @@
-﻿using EventTicketingManagementSystem.Data.Repository;
+﻿using EventTicketingManagementSystem.Constants;
+using EventTicketingManagementSystem.Data.Repository;
 using EventTicketingManagementSystem.Dtos;
 using EventTicketingManagementSystem.Models;
+using EventTicketingManagementSystem.Request;
 using EventTicketingManagementSystem.Services.Interfaces;
 
 namespace EventTicketingManagementSystem.Services.Implements
@@ -9,10 +11,12 @@ namespace EventTicketingManagementSystem.Services.Implements
     {
 
         private readonly IEventRepository _eventRepository;
+        private readonly IObjectStorageService _objectStorageService;
 
-        public EventService(IEventRepository eventRepository)
+        public EventService(IEventRepository eventRepository, IObjectStorageService objectStorageService)
         {
             _eventRepository = eventRepository;
+            _objectStorageService = objectStorageService;
         }
 
         public async Task<IEnumerable<Event>> GetAllEventsAsync()
@@ -25,24 +29,100 @@ namespace EventTicketingManagementSystem.Services.Implements
         public async Task<IEnumerable<Event>> GetEventsByFilter(string search, string category, string status) =>
             await _eventRepository.GetEventsByFilter(search, category, status);
 
-        public async Task<Event> CreateEvent(Event eventItem)
+        public async Task<int> CreateEvent(AddUpdateEventRequest eventItem)
         {
-            var eventEntity = await _eventRepository.AddAsync(eventItem);
-            await _eventRepository.SaveChangeAsync();
-            return eventEntity;
+            try
+            {
+
+                var imageUrl = string.Empty;
+                if (eventItem.Image != null)
+                {
+                    using (var fileStream = eventItem.Image.OpenReadStream())
+                    {
+                        // Generate the file name for S3 (you can modify this to meet your needs)
+                        var fileName = $"{Guid.NewGuid()}_{eventItem.Image.FileName}";
+
+                        // Call the UploadFileAsync method
+                        imageUrl = await _objectStorageService.UploadFileAsync(fileStream, fileName, CommConstants.S3_BUCKET_NAME);
+                    }
+                }
+                var eventObj = new Event
+                {
+                    Name = eventItem.Name ?? string.Empty,
+                    Description = eventItem.Description ?? string.Empty,
+                    StartDate = eventItem.StartDate.GetValueOrDefault(),
+                    EndDate = eventItem.EndDate.GetValueOrDefault(),
+                    VenueName = eventItem.VenueName,
+                    VenueAddress = eventItem.VenueAddress,
+                    ImageUrls = imageUrl,
+                    Category = eventItem.Category?.ToString(),
+                    SeatPrice = eventItem.SeatPrice
+                };
+                var eventCreated = await _eventRepository.AddAsync(eventObj);
+                await _eventRepository.SaveChangeAsync();
+
+                return eventCreated.Id;
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                throw;
+            }
+
         }
 
-        public async Task<bool> UpdateEvent(Event eventItem)
+        public async Task<bool> UpdateEvent(AddUpdateEventRequest eventItem)
         {
-            _eventRepository.Update(eventItem);
-            return await _eventRepository.SaveChangeAsync() > 0;
+            using var stream = eventItem.Image?.OpenReadStream();
+            var imageUrl = string.Empty;
+            if (stream != null)
+            {
+                imageUrl = await _objectStorageService.UploadFileAsync(stream, new Guid().ToString(), CommConstants.S3_BUCKET_NAME);
+            }
+
+            var eventObj = await _eventRepository.GetByIdAsync(eventItem.ID);
+            if (eventObj == null)
+            {
+                return false;
+            }
+
+            eventObj.Name = eventItem.Name ?? string.Empty;
+            eventObj.Description = eventItem.Description ?? string.Empty;
+            eventObj.StartDate = eventItem.StartDate.GetValueOrDefault();
+            eventObj.EndDate = eventItem.EndDate.GetValueOrDefault();
+            eventObj.VenueName = eventItem.VenueName ?? string.Empty;
+            eventObj.ImageUrls = imageUrl;
+            eventObj.VenueAddress = eventItem.VenueAddress ?? string.Empty;
+            eventObj.Category = eventItem.Category?.ToString() ?? string.Empty;
+            eventObj.SeatPrice = eventItem.SeatPrice.GetValueOrDefault();
+
+            _eventRepository.Update(eventObj);
+            var isUpdated = await _eventRepository.SaveChangeAsync() > 0;
+
+            return isUpdated;
         }
 
-        public async Task<bool> DeleteEvent(Event eventItem)
+        public async Task<bool> DeleteEvent(int id)
         {
+            var eventItem = await _eventRepository.GetByIdAsync(id);
+            
+            if (eventItem == null)
+            {
+                return false;
+            }
+
             _eventRepository.Delete(eventItem);
             return await _eventRepository.SaveChangeAsync() > 0;
         }
+
+        public async Task<IEnumerable<Event>> GetFilteredPagedEventsAsync(EventSearchParamsRequest eventFilter)
+        {
+            var events = await _eventRepository.GetFilteredPagedAsync(eventFilter);
+
+            return events;
+        }
+
+
 
 
         #region User Event
